OCREvaluator/pyproject.toml
[project]
name = "thesis2"
version = "0.1.0"
description = "OCR Accuracy Grading System using LLM for TXT files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "datasets>=4.0.0",
    "pillow>=11.3.0",
]
